requirements.txt
PyQt5
Pillow
reportlab